Rename atoms to mp4box
