+++ src/mp4box/co64.rs
@@ -1,24 +1,18 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct CttsBox {
+pub struct Co64Box {
     pub version: u8,
     pub flags: u32,
-    pub entries: Vec<CttsEntry>,
+    pub entries: Vec<u64>,
 }
 
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct CttsEntry {
-    pub sample_count: u32,
-    pub sample_offset: i32,
-}
-
-impl Mp4Box for CttsBox {
+impl Mp4Box for Co64Box {
     fn box_type() -> BoxType {
-        BoxType::CttsBox
+        BoxType::Co64Box
     }
 
     fn box_size(&self) -> u64 {
@@ -26,7 +20,7 @@ impl Mp4Box for CttsBox {
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for CttsBox {
+impl<R: Read + Seek> ReadBox<&mut R> for Co64Box {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
@@ -34,17 +28,14 @@ impl<R: Read + Seek> ReadBox<&mut R> for CttsBox {
 
         let entry_count = reader.read_u32::<BigEndian>()?;
         let mut entries = Vec::with_capacity(entry_count as usize);
-        for _ in 0..entry_count {
-            let entry = CttsEntry {
-                sample_count: reader.read_u32::<BigEndian>()?,
-                sample_offset: reader.read_i32::<BigEndian>()?,
-            };
-            entries.push(entry);
+        for _i in 0..entry_count {
+            let chunk_offset = reader.read_u64::<BigEndian>()?;
+            entries.push(chunk_offset);
         }
 
         skip_read_to(reader, start + size)?;
 
-        Ok(CttsBox {
+        Ok(Co64Box {
             version,
             flags,
             entries,
@@ -52,7 +43,7 @@ impl<R: Read + Seek> ReadBox<&mut R> for CttsBox {
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for CttsBox {
+impl<W: Write> WriteBox<&mut W> for Co64Box {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
@@ -60,9 +51,8 @@ impl<W: Write> WriteBox<&mut W> for CttsBox {
         write_box_header_ext(writer, self.version, self.flags)?;
 
         writer.write_u32::<BigEndian>(self.entries.len() as u32)?;
-        for entry in self.entries.iter() {
-            writer.write_u32::<BigEndian>(entry.sample_count)?;
-            writer.write_i32::<BigEndian>(entry.sample_offset)?;
+        for chunk_offset in self.entries.iter() {
+            writer.write_u64::<BigEndian>(*chunk_offset)?;
         }
 
         Ok(size)
@@ -72,24 +62,15 @@ impl<W: Write> WriteBox<&mut W> for CttsBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
     #[test]
-    fn test_ctts() {
-        let src_box = CttsBox {
+    fn test_co64() {
+        let src_box = Co64Box {
             version: 0,
             flags: 0,
-            entries: vec![
-                CttsEntry {
-                    sample_count: 1,
-                    sample_offset: 200,
-                },
-                CttsEntry {
-                    sample_count: 2,
-                    sample_offset: -100,
-                },
-            ],
+            entries: vec![267, 1970, 2535, 2803, 11843, 22223, 33584],
         };
         let mut buf = Vec::new();
         src_box.write_box(&mut buf).unwrap();
@@ -97,10 +78,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::CttsBox);
+        assert_eq!(header.name, BoxType::Co64Box);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = CttsBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = Co64Box::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
