+++ src/mp4box/ctts.rs
@@ -1,24 +1,24 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct SttsBox {
+pub struct CttsBox {
     pub version: u8,
     pub flags: u32,
-    pub entries: Vec<SttsEntry>,
+    pub entries: Vec<CttsEntry>,
 }
 
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct SttsEntry {
+pub struct CttsEntry {
     pub sample_count: u32,
-    pub sample_delta: u32,
+    pub sample_offset: i32,
 }
 
-impl Mp4Box for SttsBox {
+impl Mp4Box for CttsBox {
     fn box_type() -> BoxType {
-        BoxType::SttsBox
+        BoxType::CttsBox
     }
 
     fn box_size(&self) -> u64 {
@@ -26,7 +26,7 @@ impl Mp4Box for SttsBox {
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for SttsBox {
+impl<R: Read + Seek> ReadBox<&mut R> for CttsBox {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
@@ -34,17 +34,17 @@ impl<R: Read + Seek> ReadBox<&mut R> for SttsBox {
 
         let entry_count = reader.read_u32::<BigEndian>()?;
         let mut entries = Vec::with_capacity(entry_count as usize);
-        for _i in 0..entry_count {
-            let entry = SttsEntry {
+        for _ in 0..entry_count {
+            let entry = CttsEntry {
                 sample_count: reader.read_u32::<BigEndian>()?,
-                sample_delta: reader.read_u32::<BigEndian>()?,
+                sample_offset: reader.read_i32::<BigEndian>()?,
             };
             entries.push(entry);
         }
 
         skip_read_to(reader, start + size)?;
 
-        Ok(SttsBox {
+        Ok(CttsBox {
             version,
             flags,
             entries,
@@ -52,7 +52,7 @@ impl<R: Read + Seek> ReadBox<&mut R> for SttsBox {
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for SttsBox {
+impl<W: Write> WriteBox<&mut W> for CttsBox {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
@@ -62,7 +62,7 @@ impl<W: Write> WriteBox<&mut W> for SttsBox {
         writer.write_u32::<BigEndian>(self.entries.len() as u32)?;
         for entry in self.entries.iter() {
             writer.write_u32::<BigEndian>(entry.sample_count)?;
-            writer.write_u32::<BigEndian>(entry.sample_delta)?;
+            writer.write_i32::<BigEndian>(entry.sample_offset)?;
         }
 
         Ok(size)
@@ -72,22 +72,22 @@ impl<W: Write> WriteBox<&mut W> for SttsBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
     #[test]
-    fn test_stts() {
-        let src_box = SttsBox {
+    fn test_ctts() {
+        let src_box = CttsBox {
             version: 0,
             flags: 0,
             entries: vec![
-                SttsEntry {
-                    sample_count: 29726,
-                    sample_delta: 1024,
-                },
-                SttsEntry {
+                CttsEntry {
                     sample_count: 1,
-                    sample_delta: 512,
+                    sample_offset: 200,
+                },
+                CttsEntry {
+                    sample_count: 2,
+                    sample_offset: -100,
                 },
             ],
         };
@@ -97,10 +97,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::SttsBox);
+        assert_eq!(header.name, BoxType::CttsBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = SttsBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = CttsBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
