+++ src/mp4box/mvhd.rs
@@ -1,54 +1,52 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use std::char::{decode_utf16, REPLACEMENT_CHARACTER};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
 #[derive(Debug, Clone, PartialEq)]
-pub struct MdhdBox {
+pub struct MvhdBox {
     pub version: u8,
     pub flags: u32,
     pub creation_time: u64,
     pub modification_time: u64,
     pub timescale: u32,
     pub duration: u64,
-    pub language: String,
+    pub rate: FixedPointU16,
 }
 
-impl Default for MdhdBox {
+impl Default for MvhdBox {
     fn default() -> Self {
-        MdhdBox {
+        MvhdBox {
             version: 0,
             flags: 0,
             creation_time: 0,
             modification_time: 0,
             timescale: 1000,
             duration: 0,
-            language: String::from("und"),
+            rate: FixedPointU16::new(1),
         }
     }
 }
 
-impl Mp4Box for MdhdBox {
+impl Mp4Box for MvhdBox {
     fn box_type() -> BoxType {
-        BoxType::MdhdBox
+        BoxType::MvhdBox
     }
 
     fn box_size(&self) -> u64 {
         let mut size = HEADER_SIZE + HEADER_EXT_SIZE;
-
         if self.version == 1 {
             size += 28;
         } else {
             assert_eq!(self.version, 0);
             size += 16;
         }
-        size += 4;
+        size += 80;
         size
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for MdhdBox {
+impl<R: Read + Seek> ReadBox<&mut R> for MvhdBox {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
@@ -70,24 +68,23 @@ impl<R: Read + Seek> ReadBox<&mut R> for MdhdBox {
                 reader.read_u32::<BigEndian>()? as u64,
             )
         };
-        let language_code = reader.read_u16::<BigEndian>()?;
-        let language = language_string(language_code);
+        let rate = FixedPointU16::new_raw(reader.read_u32::<BigEndian>()?);
 
         skip_read_to(reader, start + size)?;
 
-        Ok(MdhdBox {
+        Ok(MvhdBox {
             version,
             flags,
             creation_time,
             modification_time,
             timescale,
             duration,
-            language,
+            rate,
         })
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for MdhdBox {
+impl<W: Write> WriteBox<&mut W> for MvhdBox {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
@@ -106,67 +103,31 @@ impl<W: Write> WriteBox<&mut W> for MdhdBox {
             writer.write_u32::<BigEndian>(self.timescale)?;
             writer.write_u32::<BigEndian>(self.duration as u32)?;
         }
+        writer.write_u32::<BigEndian>(self.rate.raw_value())?;
 
-        let language_code = language_code(&self.language);
-        writer.write_u16::<BigEndian>(language_code)?;
-        writer.write_u16::<BigEndian>(0)?; // pre-defined
+        // XXX volume, ...
+        skip_write(writer, 76)?;
 
         Ok(size)
     }
 }
 
-fn language_string(language: u16) -> String {
-    let mut lang: [u16; 3] = [0; 3];
-
-    lang[0] = ((language >> 10) & 0x1F) + 0x60;
-    lang[1] = ((language >> 5) & 0x1F) + 0x60;
-    lang[2] = ((language) & 0x1F) + 0x60;
-
-    // Decode utf-16 encoded bytes into a string.
-    let lang_str = decode_utf16(lang.iter().cloned())
-        .map(|r| r.unwrap_or(REPLACEMENT_CHARACTER))
-        .collect::<String>();
-
-    return lang_str;
-}
-
-fn language_code(language: &str) -> u16 {
-    let mut lang = language.encode_utf16();
-    let mut code = (lang.next().unwrap_or(0) & 0x1F) << 10;
-    code += (lang.next().unwrap_or(0) & 0x1F) << 5;
-    code += lang.next().unwrap_or(0) & 0x1F;
-    code
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
-    fn test_language_code(lang: &str) {
-        let code = language_code(lang);
-        let lang2 = language_string(code);
-        assert_eq!(lang, lang2);
-    }
-
     #[test]
-    fn test_language_codes() {
-        test_language_code("und");
-        test_language_code("eng");
-        test_language_code("kor");
-    }
-
-    #[test]
-    fn test_mdhd32() {
-        let src_box = MdhdBox {
+    fn test_mvhd32() {
+        let src_box = MvhdBox {
             version: 0,
             flags: 0,
             creation_time: 100,
             modification_time: 200,
-            timescale: 48000,
-            duration: 30439936,
-            language: String::from("und"),
+            timescale: 1000,
+            duration: 634634,
+            rate: FixedPointU16::new(1),
         };
         let mut buf = Vec::new();
         src_box.write_box(&mut buf).unwrap();
@@ -174,23 +135,23 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::MdhdBox);
+        assert_eq!(header.name, BoxType::MvhdBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = MdhdBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = MvhdBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 
     #[test]
-    fn test_mdhd64() {
-        let src_box = MdhdBox {
-            version: 0,
+    fn test_mvhd64() {
+        let src_box = MvhdBox {
+            version: 1,
             flags: 0,
             creation_time: 100,
             modification_time: 200,
-            timescale: 48000,
-            duration: 30439936,
-            language: String::from("eng"),
+            timescale: 1000,
+            duration: 634634,
+            rate: FixedPointU16::new(1),
         };
         let mut buf = Vec::new();
         src_box.write_box(&mut buf).unwrap();
@@ -198,10 +159,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::MdhdBox);
+        assert_eq!(header.name, BoxType::MvhdBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = MdhdBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = MvhdBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
