+++ src/mp4box/stbl.rs
@@ -1,7 +1,7 @@
 use std::io::{Read, Seek, SeekFrom, Write};
 
-use crate::atoms::*;
-use crate::atoms::{
+use crate::mp4box::*;
+use crate::mp4box::{
     co64::Co64Box, ctts::CttsBox, stco::StcoBox, stsc::StscBox, stsd::StsdBox, stss::StssBox,
     stsz::StszBox, stts::SttsBox,
 };
