+++ src/mp4box/stco.rs
@@ -1,26 +1,26 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct Co64Box {
+pub struct StcoBox {
     pub version: u8,
     pub flags: u32,
-    pub entries: Vec<u64>,
+    pub entries: Vec<u32>,
 }
 
-impl Mp4Box for Co64Box {
+impl Mp4Box for StcoBox {
     fn box_type() -> BoxType {
-        BoxType::Co64Box
+        BoxType::StcoBox
     }
 
     fn box_size(&self) -> u64 {
-        HEADER_SIZE + HEADER_EXT_SIZE + 4 + (8 * self.entries.len() as u64)
+        HEADER_SIZE + HEADER_EXT_SIZE + 4 + (4 * self.entries.len() as u64)
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for Co64Box {
+impl<R: Read + Seek> ReadBox<&mut R> for StcoBox {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
@@ -29,13 +29,13 @@ impl<R: Read + Seek> ReadBox<&mut R> for Co64Box {
         let entry_count = reader.read_u32::<BigEndian>()?;
         let mut entries = Vec::with_capacity(entry_count as usize);
         for _i in 0..entry_count {
-            let chunk_offset = reader.read_u64::<BigEndian>()?;
+            let chunk_offset = reader.read_u32::<BigEndian>()?;
             entries.push(chunk_offset);
         }
 
         skip_read_to(reader, start + size)?;
 
-        Ok(Co64Box {
+        Ok(StcoBox {
             version,
             flags,
             entries,
@@ -43,7 +43,7 @@ impl<R: Read + Seek> ReadBox<&mut R> for Co64Box {
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for Co64Box {
+impl<W: Write> WriteBox<&mut W> for StcoBox {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
@@ -52,7 +52,7 @@ impl<W: Write> WriteBox<&mut W> for Co64Box {
 
         writer.write_u32::<BigEndian>(self.entries.len() as u32)?;
         for chunk_offset in self.entries.iter() {
-            writer.write_u64::<BigEndian>(*chunk_offset)?;
+            writer.write_u32::<BigEndian>(*chunk_offset)?;
         }
 
         Ok(size)
@@ -62,12 +62,12 @@ impl<W: Write> WriteBox<&mut W> for Co64Box {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
     #[test]
-    fn test_co64() {
-        let src_box = Co64Box {
+    fn test_stco() {
+        let src_box = StcoBox {
             version: 0,
             flags: 0,
             entries: vec![267, 1970, 2535, 2803, 11843, 22223, 33584],
@@ -78,10 +78,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::Co64Box);
+        assert_eq!(header.name, BoxType::StcoBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = Co64Box::read_box(&mut reader, header.size).unwrap();
+        let dst_box = StcoBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
