+++ src/mp4box/stsd.rs
@@ -1,8 +1,8 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
-use crate::atoms::{avc1::Avc1Box, mp4a::Mp4aBox};
+use crate::mp4box::*;
+use crate::mp4box::{avc1::Avc1Box, mp4a::Mp4aBox};
 
 #[derive(Debug, Clone, PartialEq, Default)]
 pub struct StsdBox {
