+++ src/mp4box/stss.rs
@@ -1,62 +1,59 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct SmhdBox {
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct StssBox {
     pub version: u8,
     pub flags: u32,
-    pub balance: FixedPointI8,
+    pub entries: Vec<u32>,
 }
 
-impl Default for SmhdBox {
-    fn default() -> Self {
-        SmhdBox {
-            version: 0,
-            flags: 0,
-            balance: FixedPointI8::new_raw(0),
-        }
-    }
-}
-
-impl Mp4Box for SmhdBox {
+impl Mp4Box for StssBox {
     fn box_type() -> BoxType {
-        BoxType::SmhdBox
+        BoxType::StssBox
     }
 
     fn box_size(&self) -> u64 {
-        HEADER_SIZE + HEADER_EXT_SIZE + 4
+        HEADER_SIZE + HEADER_EXT_SIZE + 4 + (4 * self.entries.len() as u64)
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for SmhdBox {
+impl<R: Read + Seek> ReadBox<&mut R> for StssBox {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
         let (version, flags) = read_box_header_ext(reader)?;
 
-        let balance = FixedPointI8::new_raw(reader.read_i16::<BigEndian>()?);
+        let entry_count = reader.read_u32::<BigEndian>()?;
+        let mut entries = Vec::with_capacity(entry_count as usize);
+        for _i in 0..entry_count {
+            let sample_number = reader.read_u32::<BigEndian>()?;
+            entries.push(sample_number);
+        }
 
         skip_read_to(reader, start + size)?;
 
-        Ok(SmhdBox {
+        Ok(StssBox {
             version,
             flags,
-            balance,
+            entries,
         })
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for SmhdBox {
+impl<W: Write> WriteBox<&mut W> for StssBox {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
 
         write_box_header_ext(writer, self.version, self.flags)?;
 
-        writer.write_i16::<BigEndian>(self.balance.raw_value())?;
-        writer.write_u16::<BigEndian>(0)?; // reserved
+        writer.write_u32::<BigEndian>(self.entries.len() as u32)?;
+        for sample_number in self.entries.iter() {
+            writer.write_u32::<BigEndian>(*sample_number)?;
+        }
 
         Ok(size)
     }
@@ -65,15 +62,15 @@ impl<W: Write> WriteBox<&mut W> for SmhdBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
     #[test]
-    fn test_smhd() {
-        let src_box = SmhdBox {
+    fn test_stss() {
+        let src_box = StssBox {
             version: 0,
             flags: 0,
-            balance: FixedPointI8::new_raw(-1),
+            entries: vec![1, 61, 121, 181, 241, 301, 361, 421, 481],
         };
         let mut buf = Vec::new();
         src_box.write_box(&mut buf).unwrap();
@@ -81,10 +78,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::SmhdBox);
+        assert_eq!(header.name, BoxType::StssBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = SmhdBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = StssBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
