+++ src/mp4box/stts.rs
@@ -1,26 +1,32 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::{Read, Seek, Write};
 
-use crate::atoms::*;
+use crate::mp4box::*;
 
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct StcoBox {
+pub struct SttsBox {
     pub version: u8,
     pub flags: u32,
-    pub entries: Vec<u32>,
+    pub entries: Vec<SttsEntry>,
 }
 
-impl Mp4Box for StcoBox {
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct SttsEntry {
+    pub sample_count: u32,
+    pub sample_delta: u32,
+}
+
+impl Mp4Box for SttsBox {
     fn box_type() -> BoxType {
-        BoxType::StcoBox
+        BoxType::SttsBox
     }
 
     fn box_size(&self) -> u64 {
-        HEADER_SIZE + HEADER_EXT_SIZE + 4 + (4 * self.entries.len() as u64)
+        HEADER_SIZE + HEADER_EXT_SIZE + 4 + (8 * self.entries.len() as u64)
     }
 }
 
-impl<R: Read + Seek> ReadBox<&mut R> for StcoBox {
+impl<R: Read + Seek> ReadBox<&mut R> for SttsBox {
     fn read_box(reader: &mut R, size: u64) -> Result<Self> {
         let start = box_start(reader)?;
 
@@ -29,13 +35,16 @@ impl<R: Read + Seek> ReadBox<&mut R> for StcoBox {
         let entry_count = reader.read_u32::<BigEndian>()?;
         let mut entries = Vec::with_capacity(entry_count as usize);
         for _i in 0..entry_count {
-            let chunk_offset = reader.read_u32::<BigEndian>()?;
-            entries.push(chunk_offset);
+            let entry = SttsEntry {
+                sample_count: reader.read_u32::<BigEndian>()?,
+                sample_delta: reader.read_u32::<BigEndian>()?,
+            };
+            entries.push(entry);
         }
 
         skip_read_to(reader, start + size)?;
 
-        Ok(StcoBox {
+        Ok(SttsBox {
             version,
             flags,
             entries,
@@ -43,7 +52,7 @@ impl<R: Read + Seek> ReadBox<&mut R> for StcoBox {
     }
 }
 
-impl<W: Write> WriteBox<&mut W> for StcoBox {
+impl<W: Write> WriteBox<&mut W> for SttsBox {
     fn write_box(&self, writer: &mut W) -> Result<u64> {
         let size = self.box_size();
         BoxHeader::new(Self::box_type(), size).write(writer)?;
@@ -51,8 +60,9 @@ impl<W: Write> WriteBox<&mut W> for StcoBox {
         write_box_header_ext(writer, self.version, self.flags)?;
 
         writer.write_u32::<BigEndian>(self.entries.len() as u32)?;
-        for chunk_offset in self.entries.iter() {
-            writer.write_u32::<BigEndian>(*chunk_offset)?;
+        for entry in self.entries.iter() {
+            writer.write_u32::<BigEndian>(entry.sample_count)?;
+            writer.write_u32::<BigEndian>(entry.sample_delta)?;
         }
 
         Ok(size)
@@ -62,15 +72,24 @@ impl<W: Write> WriteBox<&mut W> for StcoBox {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::atoms::BoxHeader;
+    use crate::mp4box::BoxHeader;
     use std::io::Cursor;
 
     #[test]
-    fn test_stco() {
-        let src_box = StcoBox {
+    fn test_stts() {
+        let src_box = SttsBox {
             version: 0,
             flags: 0,
-            entries: vec![267, 1970, 2535, 2803, 11843, 22223, 33584],
+            entries: vec![
+                SttsEntry {
+                    sample_count: 29726,
+                    sample_delta: 1024,
+                },
+                SttsEntry {
+                    sample_count: 1,
+                    sample_delta: 512,
+                },
+            ],
         };
         let mut buf = Vec::new();
         src_box.write_box(&mut buf).unwrap();
@@ -78,10 +97,10 @@ mod tests {
 
         let mut reader = Cursor::new(&buf);
         let header = BoxHeader::read(&mut reader).unwrap();
-        assert_eq!(header.name, BoxType::StcoBox);
+        assert_eq!(header.name, BoxType::SttsBox);
         assert_eq!(src_box.box_size(), header.size);
 
-        let dst_box = StcoBox::read_box(&mut reader, header.size).unwrap();
+        let dst_box = SttsBox::read_box(&mut reader, header.size).unwrap();
         assert_eq!(src_box, dst_box);
     }
 }
